refactor: Playable 중 사용하지 않는 메서드 제거
